+++ tests/index.js
@@ -1587,6 +1587,73 @@ test('delete registered module', async t => {
   t.end()
 })
 
+test('delete versioned module', async t => {
+  const dir = tempy.directory()
+  const p2p = new SDK({
+    disableSwarm: true,
+    watch: false,
+    baseDir: dir
+  })
+
+  const modules = await p2p.list()
+  t.equal(modules.length, 0, 'Modules list is empty')
+
+  // create content
+  const { rawJSON: content } = await p2p.init({
+    type: 'content',
+    title: 'demo',
+    description: 'lorem ipsum'
+  })
+
+  const { rawJSON: profile } = await p2p.init({
+    type: 'profile',
+    title: 'professor X',
+    description: 'd'
+  })
+
+  const contentModules = await p2p.listContent()
+  t.equal(contentModules.length, 1, '1 content module exists')
+
+  // register
+  const authors = [encode(profile.url)]
+  // manually writing a dummy file
+  await writeFile(join(dir, encode(content.url), 'file.txt'), 'hola mundo')
+
+  const { metadata } = await p2p.set({
+    url: content.url,
+    authors,
+    main: 'file.txt'
+  })
+
+  const moduleVersioned = `${encode(content.url)}+${metadata.version}`
+  try {
+    await p2p.register(moduleVersioned, encode(profile.url))
+  } catch (err) {
+    t.fail(err.message)
+  }
+  const { rawJSON: updatedProfile } = await p2p.get(profile.url)
+  t.same(updatedProfile.contents.length, 1, 'content registered')
+
+  // hard delete
+  await p2p.delete(moduleVersioned, true)
+
+  const { rawJSON: finalProfile } = await p2p.get(profile.url)
+  t.same(finalProfile.contents.length, 0, 'content deregistered after delete')
+
+  const baseDir = await readdir(join(p2p.baseDir))
+
+  const contentModulesFinal = await p2p.listContent()
+  t.equal(contentModulesFinal.length, 0, '0 content module remains')
+
+  t.notok(
+    baseDir.includes(moduleVersioned),
+    'Module folder has been removed (deleteFiles)'
+  )
+
+  await p2p.destroy()
+  t.end()
+})
+
 test('deregister content module from profile', async t => {
   const p2p = createDb()
   const sampleContent = {