+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Fixed
 
 - listProfiles() only returned writable profiles
@@ -20,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Location of keys is now `~/.p2pcommons/.hyperdrive` instead of `~/.p2pcommons/.dat`
 - Bump module specification version to `0.9.3`
 - Bump interoperability specification version to `0.9.1`
+- Delete: allow versioned keys
 
 ### Added